refactor(web): simplify platform detection for keyboard shortcuts
Replace manual platform detection logic with a custom hook to determine if the user is on an Apple device. This change streamlines the code in both the NewChatTaskActions and DownloadButton components, enhancing maintainability and readability.
